tests/requirements.txt
numpy
pillow
pillow-avif-plugin
pixelmatch
pytest
pytest-asyncio
pytest-xprocess
seleniumbase
trame-server>=3
trame-vuetify
trame>=3.6
vtk